dashboard-service/tests/requirements-test.txt
# Test requirements for dashboard service
# Extends requirements/test.txt - includes common test dependencies
# Install with: pip install -r requirements-test.txt

# Include common test requirements
-r ../../requirements/test.txt

# Flask-specific testing
flask-testing==0.8.1

# HTML parsing for tests
beautifulsoup4==4.12.2

# Additional HTTP testing
httpretty==1.1.4

# Performance testing
pytest-benchmark==4.0.0

# Documentation testing
pytest-doctestplus==0.13.0
